clicking on search result now displays correct infoWindowContent
I know a lot of things were changed here but trust me, it is all important. I moved the class declarations for EventEmitter and Icon outside of the initMap() function, as well as some helper functions that are used in the Icon constructor. I also moved the buildings[] array (of icon objects) outside the initMap() function. All of this was necessary so that the showInfoHelper() function (which is located outside the initMap() function) can use properties of the Icon class, specifically its infoWindowContent property. This allows the showInfoHelper() function to display information about a search result when it is clicked on.